x/schwifty/keeper/msg_server_collection.go
package keeper

import (
	"context"
	"fmt"

	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/toschdev/schwifty/x/schwifty/types"
)

func (k msgServer) CreateCollection(goCtx context.Context, msg *types.MsgCreateCollection) (*types.MsgCreateCollectionResponse, error) {
	ctx := sdk.UnwrapSDKContext(goCtx)

	var collection = types.Collection{
		Creator:     msg.Creator,
		Owner:       msg.Owner,
		Name:        msg.Name,
		Description: msg.Description,
		Ticker:      msg.Ticker,
		Uri:         msg.Uri,
		UriHash:     msg.UriHash,
		Data:        msg.Data,
	}

	id := k.AppendCollection(
		ctx,
		collection,
	)

	return &types.MsgCreateCollectionResponse{
		Id: id,
	}, nil
}

func (k msgServer) UpdateCollection(goCtx context.Context, msg *types.MsgUpdateCollection) (*types.MsgUpdateCollectionResponse, error) {
	ctx := sdk.UnwrapSDKContext(goCtx)

	var collection = types.Collection{
		Creator:     msg.Creator,
		Id:          msg.Id,
		Owner:       msg.Owner,
		Name:        msg.Name,
		Description: msg.Description,
		Ticker:      msg.Ticker,
		Uri:         msg.Uri,
		UriHash:     msg.UriHash,
		Data:        msg.Data,
	}

	// Checks that the element exists
	val, found := k.GetCollection(ctx, msg.Id)
	if !found {
		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("key %d doesn't exist", msg.Id))
	}

	// Checks if the msg creator is the same as the current owner
	if msg.Creator != val.Creator {
		return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner")
	}

	k.SetCollection(ctx, collection)

	return &types.MsgUpdateCollectionResponse{}, nil
}

func (k msgServer) DeleteCollection(goCtx context.Context, msg *types.MsgDeleteCollection) (*types.MsgDeleteCollectionResponse, error) {
	ctx := sdk.UnwrapSDKContext(goCtx)

	// Checks that the element exists
	val, found := k.GetCollection(ctx, msg.Id)
	if !found {
		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("key %d doesn't exist", msg.Id))
	}

	// Checks if the msg creator is the same as the current owner
	if msg.Creator != val.Creator {
		return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner")
	}

	k.RemoveCollection(ctx, msg.Id)

	return &types.MsgDeleteCollectionResponse{}, nil
}
