x/schwifty/keeper/msg_server_collection_test.go
package keeper_test

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"github.com/toschdev/schwifty/x/schwifty/types"
)

func TestCollectionMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateCollection(ctx, &types.MsgCreateCollection{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestCollectionMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateCollection
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateCollection{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateCollection{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateCollection{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateCollection(ctx, &types.MsgCreateCollection{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateCollection(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestCollectionMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteCollection
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteCollection{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteCollection{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteCollection{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateCollection(ctx, &types.MsgCreateCollection{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteCollection(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
